Add card dialog ui

+++ app/src/main/java/com/ssoaharison/recall/helper/playback/AudioPlayer.kt
@@ -10,7 +10,7 @@ interface AudioPlayer {
     fun play()
     fun isPlaying(): Boolean
     fun hasPlayed(): Boolean
-    fun getDuration(): Float
-    fun getCurrentPosition(): Float
+    fun getDuration(): Int
+    fun getCurrentPosition(): Int
     fun onCompletion(listener: () -> Unit)
 }
